AI Based Univerity TimeTable Generator/requirements.txt
crewai
google-generativeai
pandas
openpyxl
streamlit
python-dotenv